+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Extra argument `append` for `LokiComponentExtender.add()` and `LokiComponentExtender.addMixin()` to append data after the original componentData is generated
 
 ## [2.6.1] - 14 July 2026
 ### Fixed